cloud_function/from_drive_to_storage/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage